Add language selection command
